requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
pytz==2024.2
sqlparse==0.5.3
django-cors-headers==4.6.0
drf_yasg==1.21.8
python-dotenv==1.0.1
whitenoise==6.8.2